+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Changed
+- Test against postgres versions 10, 11, 12 and 13
 
 ## [2.1.0] - 2021-03-23
 ### Added